+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 #### Improvements
 * Fix #2473: Removed unused ValidationMessages.properties
 * Fix #2408: Add documentation for Pod log options
+* Fix #2141: Decouple OpenShift Model from Kubernetes Client
 
 #### Dependency Upgrade
 * Bump Knative Serving to v0.17.2 & Knative Eventing to v0.17.3